+++ be/src/models/CounterProduct.ts
@@ -24,7 +24,7 @@ export default class CounterProduct extends Model {
   declare quantity: number;
 
   @AllowNull(false)
-  @Column(DataType.INTEGER)
+  @Column(DataType.DECIMAL(10, 2))
   declare total: number;
 
   @AllowNull(false)
