+++ ui/src/components/counters/EditionModeContent.tsx
@@ -1,0 +1,314 @@
+import React, { useState, useEffect } from 'react';
+import { Button, Grid, Typography, Paper, Box, IconButton, MenuItem, Select } from '@mui/material';
+import RemoveIcon from '@mui/icons-material/Remove';
+import AddIcon from '@mui/icons-material/Add';
+import RefreshIcon from '@mui/icons-material/Refresh';
+import CloseIcon from '@mui/icons-material/Close';
+import { DatePicker, DateValidationError, PickerChangeHandlerContext } from '@mui/x-date-pickers';
+import dayjs, { Dayjs } from 'dayjs';
+import { useAppDispatch, useAppSelector } from '../../store/hooks';
+import { fetchProducts } from '../../actions/productActions';
+import { fetchUsers } from '../../actions/userActions';
+import { Product } from '../../types/products/Product';
+import { User } from '../../types/users/User';
+import { Counter } from '../../types/counters/Counter';
+import { CounterProduct } from '../../types/counterProducts/CounterProduct';
+import { CounterUser } from '../../types/counterUsers/CounterUser';
+import { CounterProductModalProps } from '../../types/counterProducts/CounterProductModalProps';
+import { updateCounter } from '../../actions/counterActions';
+import { updateCounterProduct } from '../../actions/counterProductActions';
+import { createCounterUser, deleteCounterUser } from '../../actions/counterUserActions';
+
+const EditionModeContent: React.FC<CounterProductModalProps> = ({ table, row }) => {
+    const dispatch = useAppDispatch();
+    const { data: dataCounterProducts, loading: loadingCounterProducts, error: errorCounterProducts } = useAppSelector((state) => state.counterProducts)[0];
+    const { data: dataCounterUsers, loading: loadingCounterUsers, error: errorCounterUsers } = useAppSelector((state) => state.counterUsers)[0];
+    const { data: productData, loading: productLoading, error: productError } = useAppSelector((state) => state.products)[0];
+    const { data: userData, loading: userLoading, error: userError } = useAppSelector((state) => state.users)[0];
+    const [products, setProducts] = useState<Partial<Product>[]>([]);
+    const [selectedUserId, setSelectedUserId] = useState<number | null>(null);
+    const { loggedUserId } = useAppSelector((state) => state.session);
+
+    useEffect(() => {
+        dispatch(fetchProducts());
+        dispatch(fetchUsers());
+    }, [dispatch]);
+
+    useEffect(() => {
+        if (productData && productData.length > 0) {
+            setProducts(productData[0].data);
+        }
+    }, [productData]);
+
+    // Extract the counter ID from the row
+
+    const counterId: number | 0 = row.getValue("id");
+    const initialCounterDate: string | null = row.getValue("date");
+
+    const [counterDate, setCounterDate] = useState<Dayjs | null>(dayjs(initialCounterDate));
+
+    // Filter the counter products data based on the counterId
+    const counterProducts: Partial<CounterProduct>[] = dataCounterProducts[0]?.data.filter(
+        (product: Partial<CounterProduct>) => product.counterId === counterId
+    );
+
+    const counterUsers: Partial<CounterUser>[] = dataCounterUsers[0]?.data.filter(
+        (user: Partial<CounterUser>) => user.counterId === counterId
+    );
+
+
+    const [selectedUsers, setSelectedUsers] = useState<User[]>([]);
+    const [initialSelectedUsers, setInitialSelectedUsers] = useState<User[]>([]); // Add this line
+    
+    useEffect(() => {
+        if (userData && userData.length > 0) {
+            const userIDsInCounterUsers: number[] = counterUsers
+                .map((user) => user.userId)
+                .filter((id) => id !== undefined) as number[];
+    
+            const usersInCounterUsers: Partial<User>[] = userData[0]?.data.filter((user) =>
+                userIDsInCounterUsers.includes(user.id || 0)
+            );
+    
+            const initialUsers = usersInCounterUsers
+                .filter((user) => user.id !== undefined)
+                .map((user) => user as User);
+    
+            setSelectedUsers(initialUsers);
+            setInitialSelectedUsers([...initialUsers]); // Create a clone of selectedUsers
+        }
+    }, [userData]);
+
+    const initialQuantities: { [key: number]: number } = counterProducts.reduce((quantitiesObj, counterProduct) => {
+        const { productId, quantity } = counterProduct;
+        if (productId !== undefined && quantity !== undefined) {
+            quantitiesObj[productId] = quantity;
+        }
+        return quantitiesObj;
+    }, {} as { [key: number]: number });
+
+    const [quantities, setQuantities] = useState<{ [key: number]: number }>(initialQuantities);
+
+    const handleCancel = () => {
+        table.setEditingRow(null);
+        table.setRowSelection({});
+    };
+
+    const handleSave = async () => {
+        if (!counterDate || !selectedUsers.length || !Object.keys(quantities).length) {
+            // Add logic to handle incomplete data
+            return;
+        }
+
+        try {
+            // Step 1: Update Counter
+            const total = products.reduce((acc, product) => {
+                const id = product.id || 0;
+                const quantity = quantities[id] || 0;
+                return acc + (product.price || 0) * quantity;
+            }, 0);
+
+            const counterData: Partial<Counter> = {
+                userId: row.getValue("userId"), // For simplicity, just pick the first user
+                date: counterDate.toDate(), // Convert Dayjs to Date
+                total: total, // Calculate total
+                updatedBy: loggedUserId,
+            };
+
+            await dispatch(updateCounter({ counterId: counterId, counterData: counterData }));
+
+            // Step 2: Update Counter Products
+
+            const counterProductData: Partial<CounterProduct>[] = counterProducts.map((counterProduct) => {
+                const product: Partial<Product> = productData[0]?.data.filter(
+                    (product: Partial<Product>) => product.id === counterProduct.productId
+                )[0];
+                return {
+                    id: counterProduct.id,
+                    counterId: counterId,
+                    productId: product.id || 0,
+                    quantity: quantities[product.id || 0] || 0,
+                    total: (product.price || 0) * (quantities[product.id || 0] || 0),
+                    createdBy: loggedUserId,
+                }
+            });
+
+            await Promise.all(counterProductData.map((data) => {
+                const { id, ...rest } = data;
+                const partialCounterProduct: Partial<CounterProduct> = rest;
+                dispatch(updateCounterProduct({ counterProductId: data.id || 0, counterProductData: partialCounterProduct }))
+            }
+            ));
+
+            // Step 3: Update Counter Users
+
+            // Check if the selectedUsers list has been modified
+            const usersModified = JSON.stringify(selectedUsers) !== JSON.stringify(initialSelectedUsers);
+
+            // If the selectedUsers list has been modified, perform the necessary actions
+            if (usersModified) {
+
+                console.log("Funciona")
+                // Create an array of user IDs from selectedUsers
+                const selectedUserIds = selectedUsers.map((user) => user.id);
+
+                // Create an array of user IDs from counterUsers
+                const counterUserIds = counterUsers.map((counterUser) => counterUser.userId || 0);
+
+                // Create an array of CounterUser objects to be added
+
+                const counterUserDataToAdd: Partial<CounterUser>[] = selectedUsers.filter(
+                    (user) => !counterUserIds.includes(user.id)).map((user) => ({
+                        counterId: counterId,
+                        userId: user.id || 0,
+                        createdBy: loggedUserId,
+                    }));
+
+                // Create an array of CounterUser objects to be deleted
+
+                const counterUserDataToDelete = counterUsers.filter(
+                    (counterUser) => !selectedUserIds.includes(counterUser.userId || 0))
+                    .map((counterUser) => counterUser.id || 0);
+
+                // Dispatch actions for newly added users
+                counterUserDataToAdd.forEach((user) => dispatch(createCounterUser(user)));
+
+                // Dispatch actions for deleted users
+                counterUserDataToDelete.forEach((user) => dispatch(deleteCounterUser(user)));
+            }
+
+            // Handle success
+            table.setEditingRow(null);
+            table.setRowSelection({});
+
+        } catch (error) {
+            // Handle error
+        }
+    };
+
+    const handleDecrease = (productId: number, productTypeId: number) => {
+        setQuantities((prevQuantities) => ({
+            ...prevQuantities,
+            [productId]: Math.max((prevQuantities[productId] || 0) - (productTypeId === 2 ? -1 : 1), 0),
+        }));
+    };
+
+    const handleIncrease = (productId: number, productTypeId: number) => {
+        setQuantities((prevQuantities) => ({
+            ...prevQuantities,
+            [productId]: (prevQuantities[productId] || 0) + (productTypeId === 2 ? -1 : 1),
+        }));
+    };
+
+    const handleReset = () => {
+        setQuantities({});
+    };
+
+    const handleAddUser = () => {
+        if (selectedUserId !== null && userData) {
+            const selectedUser: Partial<User> | undefined = userData[0].data.find((user: Partial<User>) => user.id === selectedUserId);
+            if (selectedUser) {
+                // Check if the user is already selected
+                if (!selectedUsers.some((user) => user.id === selectedUser.id)) {
+                    setSelectedUsers([...selectedUsers, selectedUser as User]); // Assert selectedUser as User
+                }
+            }
+        }
+
+        // Reset selected user and selected user ID
+        setSelectedUserId(null);
+    };
+
+    const handleRemoveUser = (userId: number) => {
+        setSelectedUsers(selectedUsers.filter((user: Partial<User>) => user.id !== userId));
+    };
+
+    if (productLoading || userLoading || loadingCounterProducts || loadingCounterUsers) {
+        return <div>Loading...</div>;
+    }
+
+    if (productError || userError || errorCounterProducts || errorCounterUsers || !products.length) {
+        return <div>Error fetching data...</div>;
+    }
+
+    return <>
+        <Grid container spacing={2}>
+            {counterProducts.map((counterProduct) => {
+                const product: Partial<Product> = productData[0]?.data.filter(
+                    (product: Partial<Product>) => product.id === counterProduct.productId
+                )[0];
+                const id = product.id || 0; // Ensure id is always a number
+                const { name, price, productTypeId } = product;
+                const quantity = quantities[id] || 0;
+                const total = price ? price * quantity : 0;
+
+                return (
+                    <Grid item xs={12} key={id}>
+                        <Paper elevation={1} sx={{ p: 2 }}>
+                            <Typography variant="h6">{name}</Typography>
+                            <Box sx={{ display: 'flex', alignItems: 'center' }}>
+                                <Typography>Quantity:</Typography>
+                                <IconButton size="small" onClick={() => handleDecrease(id, productTypeId || 0)}><RemoveIcon /></IconButton>
+                                <Typography>{quantity}</Typography>
+                                <IconButton size="small" onClick={() => handleIncrease(id, productTypeId || 0)}><AddIcon /></IconButton>
+                            </Box>
+                            <Typography>Price: {price?.toLocaleString('pl-PL', { style: 'currency', currency: 'PLN' })}</Typography>
+                            <Typography>Total: {total?.toLocaleString('pl-PL', { style: 'currency', currency: 'PLN' })}</Typography>
+                        </Paper>
+                    </Grid>
+                );
+            })}
+        </Grid>
+        <Box sx={{ mt: 2, display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
+            <Typography variant="h6">Total: {products.reduce((acc, product) => {
+                const id = product.id || 0; // Ensure id is always a number
+                const quantity = quantities[id] || 0;
+                const total = product.price ? product.price * quantity : 0;
+                return acc + total;
+            }, 0)?.toLocaleString('pl-PL', { style: 'currency', currency: 'PLN' })
+            }</Typography>
+            <Button onClick={handleReset} variant="outlined" endIcon={<RefreshIcon />}>Reset</Button>
+        </Box>
+        <Box sx={{ mt: 2 }}>
+            <Typography variant="h6">Staff Working:</Typography>
+            <Paper sx={{ p: 2 }}>
+                {selectedUsers.map((user) => (
+                    <Box key={user.id} sx={{ display: 'flex', alignItems: 'center', mb: 1 }}>
+                        <Typography>{user.firstName}</Typography>
+                        <IconButton size="small" onClick={() => handleRemoveUser(user.id)}><CloseIcon /></IconButton>
+                    </Box>
+                ))}
+            </Paper>
+        </Box>
+        <Box sx={{ mt: 2 }}>
+            <Select
+                value={selectedUserId || ''}
+                onChange={(e) => setSelectedUserId(Number(e.target.value))}
+                fullWidth
+                displayEmpty
+            >
+                <MenuItem disabled value="">
+                    Select Staff
+                </MenuItem>
+                {userData && userData[0].data.map((user: Partial<User>) => (
+                    <MenuItem key={user.id} value={user.id}>{user.firstName}</MenuItem>
+                ))}
+            </Select>
+            <Button variant="outlined" onClick={handleAddUser} sx={{ mt: 2 }}>Add Staff</Button>
+        </Box>
+        <Box sx={{ mt: 2 }}>
+            <DatePicker
+                label="Counter Date"
+                format="DD/MM/YYYY"
+                value={counterDate}
+                onChange={(newValue: Dayjs | null, context: PickerChangeHandlerContext<DateValidationError>) => setCounterDate(newValue)}
+            />
+        </Box>
+        <Box sx={{ mt: 2 }}>
+            <Button variant="outlined" onClick={handleCancel} sx={{ mt: 2 }}>Cancel</Button>
+            <Button variant="contained" onClick={handleSave} sx={{ mt: 2, ml: 2 }}>Save</Button>
+        </Box>
+    </>
+}
+
+export default EditionModeContent;
