+++ ui/src/utils/ModalContent.tsx
@@ -1,5 +1,5 @@
-import { Box, Flex, Grid, Modal, ScrollArea } from "@mantine/core";
-import { MRT_EditActionButtons } from "mantine-react-table";
+import { Box, Flex, Modal, ScrollArea, Grid } from "@mantine/core";
+import { MRT_EditActionButtons, MRT_EditCellTextInput } from "mantine-react-table";
 import { useEffect } from "react";
 import { ModalContentProps } from "../types/general/ModalContentProps";
 
@@ -10,6 +10,8 @@ export const ModalContent = <T extends {}>({
     opened,
     setOpened,
     title,
+    action,
+    custom,
 }: ModalContentProps<T>) => {
     useEffect(() => {
         if (table.getState().creatingRow || table.getState().editingRow) {
@@ -22,43 +24,66 @@ export const ModalContent = <T extends {}>({
         table.setCreatingRow(null);
         table.setEditingRow(null);
     };
-    const splitIndex = Math.ceil(internalEditComponents.length / 2);
-    return (
-      <Modal
-        opened={opened}
-        onClose={handleClose}
-        centered
-        closeOnEscape 
-        closeOnClickOutside 
-        title={<div style={{ fontWeight: 'bold', fontSize: '1.25em' }}>{title}</div>}
-        scrollAreaComponent={ScrollArea.Autosize}
-        transitionProps={{ transition: 'rotate-left', duration: 200 }}
-        size="40%"
-        overlayProps={{
-          backgroundOpacity: 0.55,
-          blur: 3,
-        }}
-      >
-        <Box style={{ maxHeight: '80vh' }}>
-          {/*<Grid>
-             <Grid.Col span={6} style={{ padding: '8px 16px' }}>
-              {internalEditComponents.slice(0, splitIndex).map((field, index) => {
-                return <div key={index} style={{ marginBottom: '15px' }}>{field}</div>;
-})}
-            </Grid.Col>
-            <Grid.Col span={6} style={{ padding: '8px 16px' }}>
-              {internalEditComponents.slice(splitIndex).map((field, index) => (
-                <div key={index} style={{ marginBottom: '15px' }}>{field}</div>
-              ))}
-             
-            </Grid.Col> 
-            
-          </Grid>*/}
-          {internalEditComponents}
-          <Flex justify="flex-end">
-            <MRT_EditActionButtons row={row} table={table} variant="text" />
-          </Flex>
-        </Box>
-      </Modal>
-    );
+
+    if(custom){
+      return (
+        <Modal
+          opened={opened}
+          onClose={handleClose}
+          centered
+          closeOnEscape 
+          closeOnClickOutside 
+          title={<div style={{ fontWeight: 'bold', fontSize: '1.25em' }}>{title}</div>}
+          scrollAreaComponent={ScrollArea.Autosize}
+          transitionProps={{ transition: 'rotate-left', duration: 200 }}
+          size="40%"
+          overlayProps={{
+            backgroundOpacity: 0.55,
+            blur: 3,
+          }}
+        >
+          <Box style={{ maxHeight: '80vh' }}>
+            {internalEditComponents}
+            {/* <MRT_EditCellTextInput table = {table} cell = {internalEditComponents[1].props.cell} value={5} label={undefined}/> */}
+          </Box>
+        </Modal>
+      );
+    }else{
+      const splitIndex = Math.ceil(internalEditComponents.length / 2);
+      return (
+        <Modal
+          opened={opened}
+          onClose={handleClose}
+          centered
+          closeOnEscape 
+          closeOnClickOutside 
+          title={<div style={{ fontWeight: 'bold', fontSize: '1.25em' }}>{title}</div>}
+          scrollAreaComponent={ScrollArea.Autosize}
+          transitionProps={{ transition: 'rotate-left', duration: 200 }}
+          size="40%"
+          overlayProps={{
+            backgroundOpacity: 0.55,
+            blur: 3,
+          }}
+        >
+          <Box style={{ maxHeight: '80vh' }}>
+            <Grid>
+               <Grid.Col span={6} style={{ padding: '8px 16px' }}>
+                {internalEditComponents.slice(0, splitIndex).map((field, index) => {
+                  return <div key={index} style={{ marginBottom: '15px' }}>{field}</div>;
+                })}
+              </Grid.Col>
+              <Grid.Col span={6} style={{ padding: '8px 16px' }}>
+                {internalEditComponents.slice(splitIndex).map((field, index) => (
+                  <div key={index} style={{ marginBottom: '15px' }}>{field}</div>
+                ))}
+              </Grid.Col> 
+            </Grid>
+            <Flex justify="flex-end">
+              <MRT_EditActionButtons row={row} table={table} variant="text" />
+            </Flex>
+          </Box>
+        </Modal>
+      );
+    }
   };
